requirements.txt
# Основные зависимости
fastapi==0.104.1
streamlit==1.28.1
langchain==0.1.0
langchain-community==0.0.10
langchain-core==0.1.8
chromadb==0.4.15
ollama==0.1.7
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
uvicorn==0.24.0
python-dotenv==1.0.0

# Обработка файлов
PyMuPDF==1.23.8
python-docx==1.1.0  # Обновлено для лучшей совместимости
pandas==2.1.3
openpyxl==3.1.2
pillow==10.1.0
pytesseract==0.3.10
beautifulsoup4==4.12.2
lxml==4.9.3

# Векторизация и ML
sentence-transformers==2.2.2
transformers==4.35.2
torch>=2.0.0  # Автоматический выбор совместимой версии

# База данных
sqlalchemy==2.0.23
alembic==1.12.1

# Дополнительные
tqdm==4.66.1
jsonpath-ng==1.6.0  # Обновлено
pyyaml==6.0.1
pyinstaller==6.2.0
requests==2.31.0
aiohttp==3.9.1
tenacity==8.2.3

# Переводчики (опциональные)
# googletrans==4.0.0-rc1  # Конфликт с ollama из-за httpx
# deep-translator==1.11.4  # Альтернатива googletrans
# deepl==1.16.1  # Раскомментируйте, если нужен DeepL

# Архивы
rarfile==4.1
# zipfile36==0.1.3  # Не нужен для Python 3.12